Move ROS related landmark class into scripts folder
